+++ admin/section-list.php
@@ -22,7 +22,7 @@
         <td>{slug}</td>
         <td>{sort_order}</td>
         <td>{is_active}</td>
-        <td><a href="section-edit.php?id={id}">Edit</a> | <a href="section-delete.php?id={id}">Delete</a></td>
+        <td><a href="section-edit.php?id={id}">Edit</a> | <a href="section-delete.php?id={id}" onclick="return !!confirm('Delete this item?');">Delete</a></td>
     </tr>
     <?php
         $s = ob_get_clean();
